commmit

+++ Denis.UserList.BLL.Core/UserLogic.cs
@@ -8,10 +8,15 @@ namespace Denis.UserList.BLL.Core
     public class UserLogic : IUserLogic
     {
         private const int maxAge = 150;
+
         private readonly IUserDAO userDAO;
+        private readonly Dictionary<int, User> userCache;
+        private int LastID;
 
         public UserLogic()
         {
+            userCache = new Dictionary<int, User>();
+            LastID = 0;
             switch (Common.ReadConfigFile("user_database"))
             {
                 case "userDAO":
@@ -24,6 +29,10 @@ namespace Denis.UserList.BLL.Core
 
         public void DeleteUser(int userID)
         {
+            if (userCache.Remove(userID))
+            {
+                return;
+            }
             try
             {
                 userDAO.DeleteUser(userID);
@@ -36,17 +45,54 @@ namespace Denis.UserList.BLL.Core
 
         public IEnumerable<User> GetAllUsers()
         {
-            try
-            {
-                return userDAO.GetAllUsers();
-            }
-            catch (Exception exception)
-            {
-                throw new BLLException("Cannot get all users", exception);
-            }
+            return userCache.Values.Union(GetAllUsersFromDatabase());
         }
 
         public User GetUser(int userID)
+        {
+            if (userCache.TryGetValue(userID, out User? user))
+            {
+                return user;
+            }
+            return GetUserFromDatabase(userID);
+        }
+
+
+
+        public int AddUser(string? name, DateTime birthDate)
+        {
+            if (string.IsNullOrEmpty(name) || 
+                birthDate >= DateTime.Now || 
+                DateTimeAdditional.CompleteYearDifference(birthDate, DateTime.Now) > maxAge)
+            {
+                throw new BLLException("Incorrect user state");
+            }
+            LastID--;
+            userCache.Add(LastID, new User(LastID, name, birthDate));
+            return LastID;
+        }
+
+        public void UserAddAward(int userID, int awardID, IAwardLogic awardLogic)
+        {
+            if (userCache.TryGetValue(userID, out User? user))
+            {
+                user.AddAward(awardLogic.GetAward(awardID));
+            }    
+        }
+
+        public void ApplicationCloseEventHandler()
+        {
+            foreach (var user in userCache.Values)
+            {
+                var newUserID = AddUserToDatabase(user);
+                foreach (var award in user.GetAwards())
+                {
+                    UserAddAwardToDatabase(newUserID, award.ID);
+                }
+            }
+        }
+
+        private User GetUserFromDatabase(int userID)
         {
             try
             {
@@ -58,24 +104,7 @@ namespace Denis.UserList.BLL.Core
             }
         }
 
-        public int AddUser(string? name, DateTime birthDate)
-        {
-            if (string.IsNullOrEmpty(name) || birthDate >= DateTime.Now ||
-                DateTimeAdditional.CompleteYearDifference(birthDate, DateTime.Now) > maxAge)
-            {
-                throw new BLLException("Cannot add user");
-            }
-            try
-            {
-                return userDAO.AddUser(name, birthDate);
-            }
-            catch (Exception exception)
-            {
-                throw new BLLException("Cannot add user", exception);
-            }
-        }
-
-        public void UserAddAward(int userID, int awardID)
+        private void UserAddAwardToDatabase(int userID, int awardID)
         {
             try
             {
@@ -83,7 +112,31 @@ namespace Denis.UserList.BLL.Core
             }
             catch (Exception exception)
             {
+                throw new BLLException("Cannot add user award", exception);
+            }
+        }
+
+        private int AddUserToDatabase(User user)
+        {
+            try
+            {
+                return userDAO.AddUser(user);
+            }
+            catch (Exception exception)
+            {
                 throw new BLLException("Cannot add user", exception);
+            }
+        }
+
+        private IEnumerable<User> GetAllUsersFromDatabase()
+        {
+            try
+            {
+                return userDAO.GetAllUsers();
+            }
+            catch (Exception exception)
+            {
+                throw new BLLException("Cannot get all users", exception);
             }
         }
     }
